requirements.txt
psycopg2-binary
langchain
langchain-core
langchain-openai
